Fix discord new user auth
