+++ Dockerfile
@@ -1,0 +1,3 @@
+FROM node:20-alpine
+
+USER node
