+++ src/app/api/auth/discord/route.ts
@@ -1,20 +1,28 @@
 /**
- * Discord OAuth Login - Initiates PocketBase OAuth2 Flow
+ * Discord OAuth Login - Initiates OAuth2 Flow
  * 
- * This endpoint redirects to PocketBase's OAuth2 endpoint, which handles
- * Discord authentication directly. PocketBase manages the entire OAuth flow.
+ * This endpoint redirects to Discord's OAuth2 endpoint
  */
 import { NextRequest, NextResponse } from 'next/server';
+import { headers } from 'next/headers';
 
 export async function GET(request: NextRequest) {
-  const pbUrl = process.env.NEXT_PUBLIC_POCKETBASE_URL || 'http://localhost:8090';
-  const searchParams = request.nextUrl.searchParams;
-  const callbackUrl = searchParams.get('callbackUrl') || '/volunteer/tasks';
+  const clientId = process.env.DISCORD_CLIENT_ID!;
+  const base_url = process.env.NEXT_PUBLIC_BASE_URL!;
   
-  // Redirect to PocketBase OAuth2 endpoint
-  // PocketBase will handle the Discord OAuth flow
-  const oauthUrl = `${pbUrl}/api/oauth2-redirect?provider=discord&url=${encodeURIComponent(callbackUrl)}`;
-  
-  return NextResponse.redirect(oauthUrl);
+  const redirectUri = `${base_url}/api/auth/discord/callback`;
+
+  // Generate random state for security
+  const state = Math.random().toString(36).substring(7);
+
+  // Construct Discord OAuth URL
+  const discordAuthUrl = new URL('https://discord.com/oauth2/authorize');
+  discordAuthUrl.searchParams.set('client_id', clientId);
+  discordAuthUrl.searchParams.set('redirect_uri', redirectUri);
+  discordAuthUrl.searchParams.set('response_type', 'code');
+  discordAuthUrl.searchParams.set('scope', 'identify email');
+  discordAuthUrl.searchParams.set('state', state);
+
+  return NextResponse.redirect(discordAuthUrl.toString());
 }
